refactor(ui): move span annotation editor into it's own component
