+++ app/src/pages/trace/SpanDetails.tsx
@@ -56,6 +56,7 @@ import {
   TabList,
   Tabs,
   Text,
+  ToggleButton,
   Token,
   TokenProps,
   View,
@@ -234,6 +235,7 @@ export function SpanDetails({
     return spanHasException(span);
   }, [span]);
 
+  const [showAnnotationEditor, setShowAnnotationEditor] = useState(false);
   return (
     <PanelGroup direction="horizontal" autoSaveId="span-details-layout">
       <Panel>
@@ -271,6 +273,15 @@ export function SpanDetails({
                   size="S"
                   spanNodeId={span.id}
                   projectId={projectId}
+                />
+                <ToggleButton
+                  size="S"
+                  isSelected={showAnnotationEditor}
+                  onPress={() => {
+                    setShowAnnotationEditor(!showAnnotationEditor);
+                  }}
+                  aria-label="Toggle on the annotation editor"
+                  leadingVisual={<Icon svg={<Icons.Edit2Outline />} />}
                 />
                 <SpanActionMenu
                   traceId={span.trace.traceId}
